WATER POLLUTION DISEASE TECHNICAL ANALYSIS REPORT
Water, is a vital element of life and good health, so it is pertinent to ensure that the water for consumption is clean and safe. In this analysis, I analysed data on water pollution disease, describing the water bodies commonly used and the method of purification that different countries’ populace use on their water.